Add OnlineRecognizer
