+++ android/SherpaOnnx/app/src/main/java/com/k2fsa/sherpa/onnx/OnlineRecognizer.kt
@@ -1,0 +1,1 @@
+../../../../../../../../../../sherpa-onnx/kotlin-api/OnlineRecognizer.kt
