+++ android/SherpaOnnx/app/src/main/java/com/k2fsa/sherpa/onnx/OnlineStream.kt
@@ -1,0 +1,1 @@
+../../../../../../../../../../sherpa-onnx/kotlin-api/OnlineStream.kt
