+++ android/SherpaOnnxVadAsr/app/src/main/java/com/k2fsa/sherpa/onnx/Vad.kt
@@ -1,1 +1,1 @@
-../../../../../../../../../SherpaOnnxVad/app/src/main/java/com/k2fsa/sherpa/onnx/Vad.kt
+../../../../../../../../../../sherpa-onnx/kotlin-api/Vad.kt
