+++ kotlin-api-examples/FeatureConfig.kt
@@ -1,1 +1,1 @@
-../android/SherpaOnnxVadAsr/app/src/main/java/com/k2fsa/sherpa/onnx/FeatureConfig.kt
+../sherpa-onnx/kotlin-api/FeatureConfig.kt
