+++ kotlin-api-examples/OfflineStream.kt
@@ -1,1 +1,1 @@
-../android/SherpaOnnxAudioTagging/app/src/main/java/com/k2fsa/sherpa/onnx/audio/tagging/OfflineStream.kt
+../sherpa-onnx/kotlin-api/OfflineStream.kt
