--- kotlin-api-examples/SherpaOnnx.kt
@@ -1,1 +1,0 @@
-../android/SherpaOnnx/app/src/main/java/com/k2fsa/sherpa/onnx/SherpaOnnx.kt
