--- kotlin-api-examples/SherpaOnnx2Pass.kt
@@ -1,1 +1,0 @@
-../android/SherpaOnnx2Pass/app/src/main/java/com/k2fsa/sherpa/onnx/SherpaOnnx.kt
