+++ kotlin-api-examples/Speaker.kt
@@ -1,1 +1,1 @@
-../android/SherpaOnnxSpeakerIdentification/app/src/main/java/com/k2fsa/sherpa/onnx/speaker/identification/Speaker.kt
+../sherpa-onnx/kotlin-api/Speaker.kt
