+++ kotlin-api-examples/Vad.kt
@@ -1,1 +1,1 @@
-../android/SherpaOnnxVad/app/src/main/java/com/k2fsa/sherpa/onnx/Vad.kt
+../sherpa-onnx/kotlin-api/Vad.kt
